updates on three addition

+++ scripts/clean.js
@@ -1,7 +1,9 @@
 const sh = require('shelljs');
 const upath = require('upath');
+const fs = require('fs');
 
 const destPath = upath.resolve(upath.dirname(__filename), '../dist');
 
-sh.rm('-rf', `${destPath}/*`)
-
+if (fs.existsSync(destPath)) {
+    sh.rm('-rf', `${destPath}/*`);
+}
